requirements.txt
# ── Vector database ──────────────────────────────────────────────────────────
chromadb==0.5.3                  # local dev
qdrant-client==1.10.1            # production

# ── AI APIs ──────────────────────────────────────────────────────────────────
openai==1.40.0
anthropic==0.32.0

# ── Data collection ──────────────────────────────────────────────────────────
requests==2.32.3
aiohttp==3.9.5
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.2

# ── Data processing ──────────────────────────────────────────────────────────
pydantic==2.8.2
pydantic-settings>=2.0.0         # BaseSettings for config/settings.py
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
PyYAML==6.0.1
tqdm==4.66.4

# ── Text processing ──────────────────────────────────────────────────────────
tiktoken==0.7.0                  # token counting
nltk==3.8.1
rank-bm25==0.2.2                 # BM25 sparse retrieval
sentence-transformers==3.0.1     # cross-encoder reranking (future)

# ── Backend API ──────────────────────────────────────────────────────────────
fastapi==0.111.1
uvicorn[standard]==0.30.3
python-multipart==0.0.9          # file uploads

# ── Frontend (MVP) ────────────────────────────────────────────────────────────
streamlit==1.36.0
Pillow==10.4.0                   # image handling

# ── Utilities ────────────────────────────────────────────────────────────────
rich==13.7.1                     # pretty terminal output
loguru==0.7.2                    # logging
tenacity==8.4.2                  # retry logic
